matchd: resolve multipart msg issue with get_ports
The get_ports msg builder had two errors. First it would not
correctly cancel messages when the buffer was overrun resulting
in parse errors on the receiving side. And second it could skip
ports because the counter was advancing even when the error
indicating it needed to retry the msg packing was given.

This patch resolves these issues.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Acked-by: Jeff Shaw <jeffrey.b.shaw@intel.com>