Add match_nl_set_verbose to suppress rule logging
matchlib_nl has global verbose variable it passes to match library
calls. However, it did not pass it to an explicit pp_rule() call in
match_nl_set_del_rules(). This patch fixes that and add a function that
allows users of matchlib_nl to set verbose in matchlib_nl to a desirable
value. It also makes verbose module level instead of global.

Signed-off-by: Ronen Arad <ronen.arad@intel.com>
Acked-by: Jeff Shaw <jeffrey.b.shaw@intel.com>