ieslib: add pcie ports as untagged members of vlans
Here we assume that if an Ethernet port is being added as
a member of the VLAN, it should be a tagged member so frames egress
the port on the vlan with the tag inserted.

This is not what we want for host-interface (aka pcie) ports since
the vlan info is passed through the descriptor.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Sushma Sitaram <sushma.sitaram@intel.com>