ieslib: configure tunnel engine table to set sglort
Previously, when frames were forwarded from the tunnel engine
they retained the original source glort.  When matching the
ingress logical port, it was not possible to distinguish between
frames which ingressed on a logical port vs. frames which were
forwarded from the tunnel which originally ingressed on a logical
port.  This commit configures new tunnel engine tables such that
the tunnel engine will write a default source glort instead of
passing the original through.

Fixes: ONP-2548

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Acked-by: Hao Zheng <hao.zheng@intel.com>