matchd_lib: fix match mask check
When a table is created, each header_instance.header_field needs
to have a mask specified, and the mask has to be compatible with
the mask type supported for the header.field in the source table.

Previously, only the first header instance was being checked.  This
commit fixes the behavior so that all header.field mask types are
checked against the source table.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>