ieslib: limit get/set ports to valid ports
Previously, a 'get_ports' command will return most of the ports
configured in the hardware, regardless of whether or not they
are enabled, or configurable.

This commit makes it so only configurable ports are returned by
a get_ports command, and that a set_port command can only modify
port settings for configurable ports. Note that not all port
settings are write-able for all port types.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>