matchd: Refactor netlink recv in matchd and libmatchd
This patch replaces the receive loop within matchd.c:main() with
a call to a new function matchd_receive_loop() within libmatchd.
The objective is to replace nl_recv() with nl_recvmsgs_default()
and thus allow for protocol debugging of matchd using NLCB=debug.
The logic of matchd_receive_loop() is similar to the logic some
clients already use.

Signed-off-by: Ronen Arad <ronen.arad@intel.com>